rename _key.py to _encryption.py and update imports
